Feature: monolith mode (#2)
* feat: Obtain public key of a order

Added the logic for search the public key of a order.

* fix: get public key

Fixed error when obtaining the public key to continue with the validation process.

* feat: verifying binance signatre

The payment is verified to know if it is legitimate or not.

* feat: added console logs

* update: Monolith mode

The project is organized in such a way that the front-end and back-end are separated.

* feat: init ionic